pilot: Extract world render logic into RWorldRenderer2d (issue #754 PR1)
Move R2DWidget::paintWorld into a new RWorldRenderer2d class in
RWorldRenderer2d.{hpp,cpp}. R2DWidget::paintEvent constructs
RWorldRenderer2d{world, view} and calls paint(painter), keeping
the on-screen widget and the future offscreen renderer on one code path.
World::points() is made const to allow the renderer to hold a const
world reference. Pure refactor; rendered pixels unchanged.
